Util folders for HR and SPA
