+++ tests/HRTests/utils/enumerateSMs.py
