+++ tests/SPATests/utils/enumerateSMs.py
