+++ tests/SPATests/utils/minmaxSMs.py
